requirements.txt
streamlit==1.36.0
tensorflow-cpu==2.13.0
keras==2.13.1
numpy==1.24.4